fixed alleged type mismatch errors (hopefully)

+++ architecture_simulator/uarch/architectural_state.py
@@ -19,27 +19,37 @@ class Memory:
 
     def load_halfword(self, adress: int) -> fixedint.MutableUInt16:
         return (
-            fixedint.MutableUInt16(self.memory_file[adress + 1 % pow(2, 32)]) << 8
+            fixedint.MutableUInt16(int(self.memory_file[adress + 1 % pow(2, 32)])) << 8
             | self.memory_file[adress % pow(2, 32)]
         )
 
     def store_halfword(self, adress: int, value: fixedint.MutableUInt16):
-        self.memory_file[adress % pow(2, 32)] = fixedint.MutableUInt8(value[0:8])
-        self.memory_file[adress + 1 % pow(2, 32)] = fixedint.MutableUInt8(value[8:16])
+        self.memory_file[adress % pow(2, 32)] = fixedint.MutableUInt8(int(value[0:8]))
+        self.memory_file[adress + 1 % pow(2, 32)] = fixedint.MutableUInt8(
+            int(value[8:16])
+        )
 
     def load_word(self, adress: int) -> fixedint.MutableUInt32:
         return (
-            fixedint.MutableUInt32(self.memory_file[adress + 3 % pow(2, 32)]) << 24
-            | fixedint.MutableUInt32(self.memory_file[adress + 2 % pow(2, 32)]) << 16
-            | fixedint.MutableUInt32(self.memory_file[adress + 1 % pow(2, 32)]) << 8
-            | fixedint.MutableUInt32(self.memory_file[adress % pow(2, 32)])
+            fixedint.MutableUInt32(int(self.memory_file[adress + 3 % pow(2, 32)])) << 24
+            | fixedint.MutableUInt32(int(self.memory_file[adress + 2 % pow(2, 32)]))
+            << 16
+            | fixedint.MutableUInt32(int(self.memory_file[adress + 1 % pow(2, 32)]))
+            << 8
+            | fixedint.MutableUInt32(int(self.memory_file[adress % pow(2, 32)]))
         )
 
     def store_word(self, adress: int, value: fixedint.MutableUInt32):
-        self.memory_file[adress % pow(2, 32)] = fixedint.MutableUInt8(value[0:8])
-        self.memory_file[adress + 1 % pow(2, 32)] = fixedint.MutableUInt8(value[8:16])
-        self.memory_file[adress + 2 % pow(2, 32)] = fixedint.MutableUInt8(value[16:24])
-        self.memory_file[adress + 3 % pow(2, 32)] = fixedint.MutableUInt8(value[24:32])
+        self.memory_file[adress % pow(2, 32)] = fixedint.MutableUInt8(int(value[0:8]))
+        self.memory_file[adress + 1 % pow(2, 32)] = fixedint.MutableUInt8(
+            int(value[8:16])
+        )
+        self.memory_file[adress + 2 % pow(2, 32)] = fixedint.MutableUInt8(
+            int(value[16:24])
+        )
+        self.memory_file[adress + 3 % pow(2, 32)] = fixedint.MutableUInt8(
+            int(value[24:32])
+        )
 
 
 @dataclass
